Written dialogues for corridor, but not implemented WIP

+++ Assets/Scripts/Skill System/AddSkillPointsNode.cs
@@ -26,7 +26,7 @@ public class AddSkillPointsNode : Unit
 
         for (int i = 0; i < skillCount; i++)
         {
-            amountInputs.Add(ValueInput<uint>($"Amount {i + 1}", 0));
+            amountInputs.Add(ValueInput<float>($"Amount {i + 1}", 0f));
             skillInputs.Add(ValueInput<SkillSO>($"Skill {i + 1}", null));
         }
 
@@ -39,7 +39,7 @@ public class AddSkillPointsNode : Unit
 
         for (int i = 0; i < skillCount; i++)
         {
-            int amount = (int)flow.GetValue<uint>(amountInputs[i]);
+            float amount = flow.GetValue<float>(amountInputs[i]);
             SkillSO skill = flow.GetValue<SkillSO>(skillInputs[i]);
 
             if (skill != null)
